fix gather script (again)

+++ gather.ps1
@@ -12,9 +12,14 @@ Function Find-NewName {
     $Path
 }
 
-mkdir .\all
-rm -r .\all\*
+if (Test-Path .\all) {
+    rm -r .\all\*
+} else {
+    mkdir .\all
+}
+
 git submodule sync
-git submodule update
+git submodule init
+git submodule update --remote
 Get-ChildItem .\TCL\docs\* -Include *.md -Exclude _* -Recurse | % { Copy-Item -Path $_ -Destination (Find-NewName ".\all\$((Get-Item $_).Name)") }
 python clean.py all
